% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ratio.R
\name{absz.density.ratio}
\alias{absz.density.ratio}
\title{Perform kernel estimates of two densities of absolute z-statistics
and their ratio.}
\usage{
absz.density.ratio(
  z.num,
  z.denom,
  at,
  bootstrap = TRUE,
  B = 1000,
  ci.level = 0.95,
  bw = 0.1,
  kernel = "epanechnikov",
  return.as = c("long", "wide")[1],
  ...
)
}
\arguments{
\item{z.num}{observed z-statistics forming numerator density}

\item{z.denom}{observed z-statistics forming denominator density}

\item{at}{position where density shall be evaluated}

\item{bootstrap}{if TRUE add bootstrap SE and CI for all measures}

\item{B}{number of bootstrap repetitions}

\item{ci.level}{Confidence level. Default 0.95.}

\item{...}{arguments for absz.density}
}
\description{
Add by default bootstrap standard errors and confidence intervals
}
