% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_absz.R
\name{absz.density}
\alias{absz.density}
\title{Density estimates for absolute z-statistics assuming that z-statistics are symmetrically distributed around 0}
\usage{
absz.density(
  z,
  at = NULL,
  bw = 0.1,
  adjust = 1,
  kernel = "epanechnikov",
  n = 1024
)
}
\arguments{
\item{z}{vector of z-statistics (or absolute z-statistics)}

\item{at}{vector of points where density shall be evaluated. If NULL return a function (by calling \code{approxfun}) that allows evaluate the density at arbitrary points.}

\item{bw, adjust, kernel, n, ...}{arguments passed to \code{stats::density}}
}
\description{
Avoids downward bias at the left hand side where abs(z)=0.
}
