% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{as.perc}
\alias{as.perc}
\title{Convert numbers like 0.421 to 42.1\%}
\usage{
as.perc(x, digits = 1)
}
\arguments{
\item{x}{a vector of floating point numbers}

\item{digits}{to how many decimal digits shall the percentage be rounded?}
}
\description{
Convert numbers like 0.421 to 42.1\%
}
