% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derounding.R
\name{compute.theta.stats}
\alias{compute.theta.stats}
\title{theta shall be the probability that a z-statistic is
above a threshold z0 in a window with half-width h around z0}
\usage{
\method{compute}{theta.stats}(z, z0 = 1.96, h = NULL)
}
\arguments{
\item{z}{A vector of z-values, possible a derounding draw}

\item{z0}{The critical treshold of z that marks an observation significant. This can be a single value, or a vector of the same length as z. This allows to use critical t-values that are adjusted for the sample size of the test an observation refers to}

\item{h}{The half-width of the considered window around z0. If NULL assume that all observations are already in the correct window.}
}
\description{
This function returns for a given vector of z
(possibly one derounding draw) statistics:
estimate, standard error, 95% CI of theta
and number of observations etc
}
\details{
This can be used inside Monte Carlo simulations
}
