% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deround_dsr.R
\name{dsr.ab.df}
\alias{dsr.ab.df}
\title{Create an ab.df for the dsr approach}
\usage{
dsr.ab.df(
  dat,
  h.seq = c(0.05, 0.075, 0.1, 0.2, 0.3, 0.4, 0.5),
  z0 = 1.96,
  min.n = 10000,
  min.rounds = 5,
  verbose = TRUE
)
}
\arguments{
\item{dat}{the data frame that should contain at least the columns z and num.deci (number fo decimal places of mu and sigma, maximum of both)}

\item{h.seq}{all considered window half-widths}

\item{z0}{the significance threshold. Can be a single number or a vector with one element per row of dat.}

\item{min.n}{how many z values shall be minimally rounded to compute the derounded z-distribution for each observation.}

\item{min.rounds}{how many repetitions of rounding z-values shall there be at least (even if min.n is already reached).}

\item{verbose}{Shall some progress information be shown? (This function can take a while).}
}
\description{
The resulting data frame is required for derounding b simulting rounding (dsr) approach.
It contains a row for all considered combinations of z and s and window half-width h in h.seq. The columns share.below and share.above indicate which share of derounded z-statistics are inside the window and either fall below or above the threshold z0, respectively. Note that 1-share.above-share.below is the share of derounded z-statistics that fall outside the considered window.
}
