% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deround_zda.R
\name{make.z.pdf}
\alias{make.z.pdf}
\title{Compute a normalized pdf from a vector of z-statistics}
\usage{
make.z.pdf(
  z,
  bw = 0.02,
  kernel = "epanechnikov",
  n = 512,
  dat,
  min.s = 100,
  z.min = 0,
  z.max = 5,
  show.hist = FALSE,
  ...
)
}
\arguments{
\item{z}{a vector of z statistics. Usually, you would select
all values from dat whose mu and sigma have sufficiently
many significant digits}

\item{...}{other parameters passed to \code{\link{stats::density}}}
}
\description{
The PDF is normalized such that the point of highest density is 1
}
