% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{num.deci}
\alias{num.deci}
\title{Get the number of significand digits of a floating point number
using the character presentation of those numbers of R}
\usage{
num.deci(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Get the number of significand digits of a floating point number
using the character presentation of those numbers of R
}
