% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{num.sig.digits}
\alias{num.sig.digits}
\title{Get the number of significand digits of a floating point number
using the character presentation of those numbers of R}
\usage{
num.sig.digits(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
We assume that trailing zeros left of the decimal point are
significant digits while trailing zeros right of the decimal point
are not significant digits
}
