% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedprobit.R
\name{probitmfxweights}
\alias{probitmfxweights}
\title{Weighted probit with marginal effects and cluster robust se}
\usage{
probitmfxweights(
  formula,
  data,
  atmean = TRUE,
  robust = FALSE,
  clustervar1 = NULL,
  clustervar2 = NULL,
  start = NULL,
  control = list(),
  weights = NULL
)
}
\description{
Weighted probit with marginal effects and cluster robust se
}
