% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedprobit.R
\name{probitweighted}
\alias{probitweighted}
\title{Function to estimate clustered weighted probit regression with clustered standard errors}
\usage{
probitweighted(
  formula,
  data,
  atmean = TRUE,
  robust = FALSE,
  clustervar1 = NULL,
  clustervar2 = NULL,
  start = NULL,
  control = list(),
  weights = NULL
)
}
\description{
Function to estimate clustered weighted probit regression with clustered standard errors
}
