% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{rightmost.sig.digit}
\alias{rightmost.sig.digit}
\title{Get the last significant digit(s) of a floating point number}
\usage{
rightmost.sig.digit(x, r1 = 1, r2 = 1)
}
\arguments{
\item{x}{The vector of floating point numbers}

\item{r1}{Starting position from right}

\item{r2}{Ending position from right}
}
\description{
Get the last significant digit(s) of a floating point number
}
