% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\name{rounding.risk.s.thresholds}
\alias{rounding.risk.s.thresholds}
\title{Compute thresholds for the significant s of the
reported standard deviation such that we can rule-out
the errors: misclassification, wrong inclusion, wrong exclusion}
\usage{
rounding.risk.s.thresholds(z, z0 = z0, h = 0.2)
}
\arguments{
\item{z}{a vector of z statistics}

\item{z0}{significance threshold. Can be a single number or a vector of length z}

\item{h}{half-width of considered window around z0}
}
\value{
A data frame with the columns "z", "s.misclass", "s.include", "s.exclude" specifying for each z value the corresponding thresholds.
}
\description{
Compute thresholds for the significant s of the
reported standard deviation such that we can rule-out
the errors: misclassification, wrong inclusion, wrong exclusion
}
