% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\name{rounding.risks}
\alias{rounding.risks}
\title{Assess for observations with reported z-statistic z and
a signficand of s for the standard error whether it is at risk of
the errors:
misclassification, wrong inclusion, wrong exclusion}
\usage{
rounding.risks(z, s, z0 = 1.96, h = 0.2)
}
\arguments{
\item{z}{a vector of z statistics}

\item{s}{vector of corresponding significands of the standard error}

\item{z0}{significance threshold. Can be a single number like 1.96 or a vector of length z}

\item{h}{half-width of considered window around z0}
}
\value{
A data frame with risk of missclassification information for each observations. We illustrate the columns for the misclassification risk:
 "s.misclass" is the threshold for the significand s above which we can rule out misclassification risk
 \code{risk.misclass = s < s.misclass} indicates whether the observation is at risk of misclassification
  \code{risk.misclass.below = risk.misclass & z < z0} indicates whether the observation is at risk of misclassification and below the significance threshold
  the other columns should be self-explainable given this info.
}
\description{
Assess for observations with reported z-statistic z and
a signficand of s for the standard error whether it is at risk of
the errors:
misclassification, wrong inclusion, wrong exclusion
}
