% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\name{rounding.risks.summary}
\alias{rounding.risks.summary}
\title{Summary statistics for rounding risks for different thresholds}
\usage{
rounding.risks.summary(rr.dat, s.thresh = 0:100, long = TRUE)
}
\arguments{
\item{rr.dat}{A data frame returned from a call to \code{rounding.risks}.}

\item{long}{if TRUE (default return results in a long format)}

\item{s.tresh}{a vector of considered s thresholds}
}
\description{
Summary statistics for rounding risks for different thresholds
}
