% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\name{sample.uniform.z.deround}
\alias{sample.uniform.z.deround}
\title{Sample derounded z from the uniformely derounded distributon for a given
single value of mu and sigma}
\usage{
sample.uniform.z.deround(
  n,
  mu,
  sigma,
  mu.dec = pmax(num.deci(mu), num.deci(sigma)),
  sigma.dec = mu.dec
)
}
\arguments{
\item{n}{Number of sample draws}

\item{mu}{Reported coefficient, possibly rounded}

\item{sigma}{Reported standard error, possibly rounded.}

\item{mu.dec}{Number of decimal places mu is reported to. Usually, we would assume that mu and sigma are rounded to the same number of decimal places. Since trailing zeros may not be detected, we set the default \code{mu.dec=pmax(num.deci(mu.round),num.deci(sigma.round))}.}

\item{sigma.dec}{By default equal to mu.dec.}
}
\description{
Sample derounded z from the uniformely derounded distributon for a given
single value of mu and sigma
}
