% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{set.last.digit.zero}
\alias{set.last.digit.zero}
\title{Sets the last digit of a number x to zero}
\usage{
set.last.digit.zero(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Sets the last digit of a number x to zero
}
