% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_funs.R
\name{window.binom.test.2s}
\alias{window.binom.test.2s}
\title{Apply on windows two sided binomiminal test with H0: z = z0}
\usage{
\method{window}{binom.test.2s}(above = z >= z0, h = NA, ci.level = 0.9, z, z0, ...)
}
\description{
Apply on windows two sided binomiminal test with H0: z = z0
}
