% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_funs.R
\name{window.binom.test}
\alias{window.binom.test}
\title{Apply on windows one-sided binomiminal test with H0: z <= z0}
\usage{
\method{window}{binom.test}(above = z >= z0, h = NA, ci.level = 0.95, z, z0, ...)
}
\description{
Apply on windows one-sided binomiminal test with H0: z <= z0
}
