% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_funs.R
\name{window.t.ci}
\alias{window.t.ci}
\title{Window function returning estimated probability that a z-statistic is
above a threshold z0 in a window with half-width h around z0 and t-test
confidence intervals}
\usage{
\method{window}{t.ci}(above = z >= z0, h = NA, ci.level = 0.95, z, z0, ...)
}
\description{
Can be used as argument \code{window.fun} in \code{\link{compute.with.derounding}}
}
