% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deround.R
\name{deround.z.uniform}
\alias{deround.z.uniform}
\title{Draw derounded z assuming missing digits of mu and sigma are uniformly distributed}
\usage{
deround.z.uniform(
  mu,
  sigma,
  mu.dec = pmax(num.deci(mu), num.deci(sigma)),
  sigma.dec = mu.dec
)
}
\arguments{
\item{mu}{Reported coefficient, possibly rounded}

\item{sigma}{Reported standard error, possibly rounded.}

\item{mu.dec}{Number of decimal places mu is reported to. Usually, we would assume that mu and sigma are rounded to the same number of decimal places. Since trailing zeros may not be detected, we set the default \code{mu.dec=pmax(num.deci(mu),num.deci(sigma))}.}

\item{sigma.dec}{By default equal to mu.dec.}
}
\description{
Draw derounded z assuming missing digits of mu and sigma are uniformly distributed
}
