% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_tools.R
\name{significand}
\alias{significand}
\title{Get the significands of a numeric vector using
the character presentation of those numbers of R}
\usage{
significand(x, num.deci = NULL)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
The significand is the integer of all significand digits, e.g.
the significand of 0.012 is 12.
}
